% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uHMT_download.R
\name{uHMT_download}
\alias{uHMT_download}
\title{A uHMT Get Data Function}
\usage{
uHMT_download(omit.siten = c(2:19))
}
\arguments{
\item{omit.siten}{Which stations (identified by station number) would you like to not download. Defaults to all sites except the first one. Thus only site 1 data will download.}
}
\description{
This function downloads uHMT data files from the NOAA CESSRST online database.
}
\examples{
## The below expression will download all uHMT station's data except station numbers 4, 17, and 19
uhmt_ls <- uHMT_download(omit.siten = c(4,17,19))
}
