% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uHMT_write_csv.R
\name{uHMT_write_csv}
\alias{uHMT_write_csv}
\title{A uHMT Data Exporting Function}
\usage{
uHMT_write_csv(lsofuhmt, filename = NULL)
}
\arguments{
\item{lsofuhmt}{a list containing one data frame per station.}

\item{filename}{a string to assign to the beginning of the csv filename. This arguement is optional, where the default is "uhmt_csv_stationname.csv".}
}
\description{
This function writes out uHMT data files as multiple csv's, depending on the number of objects in the provided list. It is usually used to write out csv files indvidually for each station.
}
